pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sfmc-client"
version = "0.1.0"
description = "A lightweight Python client for Salesforce Marketing Cloud (SFMC) with modular object managers for common API operations."
authors = [
    { name = "Bobby Curley", email = "bcurley.dev@gmail.com" }
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
readme = "README.md"
requires-python = ">=3.10.5"

dependencies = [
    "aiohttp>=3.12.13,<4.0.0",
    "pipdeptree>=2.27.0,<3.0.0",
    "pytest-cov>=6.2.1,<7.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "requests>=2.32.3,<3.0.0",
    "setuptools>=58.1.0,<59.0.0"
]

[project.urls]
Homepage = "https://github.com/curleyr/sfmc-client-python"
Repository = "https://github.com/curleyr/sfmc-client-python"
Issues = "https://github.com/curleyr/sfmc-client-python/issues"

[tool.setuptools.packages.find]
where = ["sfmc_client"]